Refactor : change transaction amount to BigDecimal

+++ src/test/java/features/DepositMoneyFeature.java
@@ -6,6 +6,8 @@ import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class DepositMoneyFeature {
@@ -21,12 +23,12 @@ public class DepositMoneyFeature {
     }
 
     @When("^he deposits \"([^\"]*)\" EUR into his account$")
-    public void depositMoneyIntoAccount(double amount) throws Throwable {
+    public void depositMoneyIntoAccount(BigDecimal amount) throws Throwable {
         account.deposit(amount);
     }
 
     @Then("^his new balance is \"([^\"]*)\" EUR$")
-    public void verifyRunningBalance(double amount) throws Throwable {
+    public void verifyRunningBalance(BigDecimal amount) throws Throwable {
         assertThat(account.runningBalance()).isEqualTo(amount);
     }
